Fixes tables and schema

+++ HW07-spring-data/src/main/java/ru/otus/hw07/dao/AuthorRepository.java
@@ -3,7 +3,7 @@ package ru.otus.hw07.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.otus.hw07.domain.Author;
 
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 
     Author getByFirstNameAndLastName(String firstName, String lastName);
 }
